refactor: photoIdentifier 변환 로직 및 권한 처리 서비스 분리
- PhotoIdentifier 변환 로직을 별도 서비스로 분리
  - toPhotoIdentifier: URI → PhotoIdentifier 변환
  - toPhotoIdentifierFromImage: ImagePicker → PhotoIdentifier 변환
- 카메라 권한 로직을 permission.hook.ts로 통합
- MediaPickerBottomSheet 중복 코드 제거 (70줄 감소)
- hero.type.ts에서 비즈니스 로직 제거, 타입 정의에만 집중
- 관심사 분리 및 코드 재사용성 향상
